Added support for multiple segments and included data storage
A cell can be intersected by multiple well segments, all defined by an
entry point and an exit point. Each segment can have a different radius.

A static method will help in identifying whether a cell is already known
or not based on the global index of the corresponding grid cell

Added a data storage to save intermediate caluclation results.
Renamed methods to make explicit if it targets a segment or the cell.