Added WellDefinition class, added code that works with multiple wells with multiple segments
This class is used to store a well that is defined as a list of
segments. Each segment is defined by a pair of <heel, toe> and a radius.
A well definition can be read from a file. The ReadFromFile is a static
class method which will read multiple wells from the file and create
instances of WellDefinition in a vector.

The new code calculates the well index for all block that are
intersected by a segment. The calculation is done per well so a block
intersected by multiple wells will appear multiple times in the compdat
section for each well.
Debug information is saved for each block containing the projection
method quantities.