Commented out intersection functions that need more attention
The main isssue with this tool is that the search for cells intersected
by well segments is too slow, especially for segments that are outside
or the grid or intersect inactive cells. A better/faster way to do this
is needed.

The fucntions GetIntersection, InBox and CheckLineBox where part of a
intial trial to do just that but in the previous version of the tool
were only used to test for segments that are completely outside the
reservoir. They have been commented out and replaced with a new method
IsLineCompletelyOutsideBox that tests only if the line is completely
outside the box.

The method collect_intersected_cells is still considered as "work in
progress" and, therefore,  contains quite some commented out debug
statements and it is not optimally written.