Change how the line is traversed when finding new end points.
Instead of always taking very small steps, the methods now takes steps with length equal to one tenth of the smallest dimension in any grid cell until it finds itself inside a cell. Then, it takes steps (with the old small step length) until it is back outside the cell, and returns the last point that was inside a cell.

This change seems to fix issues we had with calculations taking a very long time for some paths. The method has been tested for the problematic paths listed in #12 which now all finish quickly. The new method will be tested for other paths in other reservoirs that have proved problematic.