Rewrote the procedure for locating intersected cells.
This new implementation solves the issue of cycling between blocks, and should, overall, be more robust.

The new implementation is a hybrid of Hilmar's original implementation for locating the next block and Alin's method for stepping into the reservoir when we're outside.

We now step along the spline, fraction by fraction, and stop when the `step` variable is greater than 1.0. This is the same as what we do to step into the reservoir to find new heels and toes if needed. The difference is that here, we only use it to step from the exit point in one block through to the entry point in the next point. After that, the old `find_exit_point` method is used to find the exit point, and `step` is updated accordingly. Stepping along the well in this way instead of using the old method for getting into the next block eliminates the issue of cycling between blocks.

Note that this new version will sometimes not find blocks where the line only passes through a corner (i.e. it touches the block, but it does not actually pass through it). These blocks were found by the old implementation, but as they will get well indices approximately (or exactly) equal to zero, not including them should have little to no effect when it comes to the actual simulations.