Fix cycling issue.
The issue where the WIC would cycle for paths through strange cells in the Norne reservoir has been fixed. The issue was caused by the find_exit_point_method sometimes not finding the correct exit point.

When locating the intersected cells, we now check whether each cell we add will introduce a cycle. If it will, we switch to traversing through the last added cell to find a new exit point, and then passing through to the next cell.