Fix bug with inactive cells
Calculating the intersection of a well segment with the grid blocks was wrong after an inactive cell was encountered.
The calculation of the intersection point used the last exit point as "exception point". This was wrong because the last exit point could have been many inactive cells away on the well segment trajectory.
Also the entry point used to be set at the first point inside a cell and not calculated exactly. This resulted in inacurate  projections, usually understimating the projections.

The exact entry point is calculated by finding the exit point backwords.

The epsilon has  been changed to be the 1/100 of the smallest cell dimension.